incremental scheduler: decompose topological sort of clusters
After incrementally combining SCCs into clusters,
the remaining clusters are topologically sorted and
combined in a sequence node.
This is always correct, but it may be too strict in some cases.
In particular, while the clustering is being performed
inside a connected component of the dependence graph,
some of the clusters may still be pairwise independent and
the sequence imposes an arbitrary order on them.
Try and decompose the topological sort of the clusters
to expose some of this freedom by also introducing set nodes.

The algorithm recursively isolates a single cluster and
then detects independent components among the clusters
before the selected cluster (in the topological order) and
also among the clusters after the selected cluster.
When components are detected in the first and/or last piece,
a set node is introduced to split them up.
The three pieces are then combined into a sequence node.
For simplicity, no cluster is allowed to move across
the selected cluster, even if this may be allowed
by the schedule constraints.

The cluster selected to isolate is the one with
the highest number of incoming and outgoing edges.
Isolating the cluster cuts these edges, increasing
the chance of detecting components in the resulting two pieces.
If the clusters form a chain then there can be no further
independent components and the recursion is terminated.

Requested-by: Nana CK <1429802329@qq.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>